Bug fixes
* Fix bug in writeFile function where files were written asynchronously
* Fix bug in PostCSS Email Dark Mode plugin where media queries with multiple rules weren't formatted correctly
* Fix bug with Prettier formatting where htmlwhitespace was ignored
* Fix bug in logger where logs would fail if there wasn't a spinner present
* Move from node-glob to globby
* Tidying up files and removing unused code
