+++ bin/functions/errors.js
@@ -1,4 +1,4 @@
-import path from 'path';
+
 import fs from 'fs';
 
 import { log } from './logger.js';
