+++ bin/functions/getdata.js
@@ -34,62 +34,6 @@ const getFilepaths = (baseDir, type) => {
   return data;
 };
 
-
-class getSassData {
-    constructor(options) {
-      this.canonicalize = this.canonicalize.bind(this);
-      this.load = this.load.bind(this);
-
-      this.dataDir = (options && options.dataDir) ? options.dataDir : '';
-
-    }
-    canonicalize(url) {
-        if (!url.endsWith('.json')) return null;
-
-        if (url.startsWith('https//') || url.startsWith('http://')) {
-          return null
-        }
-
-        const dataDir = (url.startsWith('./')) ? '' : this.dataDir;
-          
-        const filePath = path.resolve(dataDir, url);
-
-        if (!fs.existsSync(filePath)) return null;
-
-        return new URL(`file://${filePath}`);
-        
-    }
-    load(canonicalUrl) {
-
-        let jsonString = fs.readFileSync(canonicalUrl, 'utf-8');
-
-        // read the contents of the JSON file
-        const jsonData = JSON.parse(jsonString);
-
-        // convert the JSON data to SASS variables
-        let sassVars = '';
-        function parseJson(data, parentKey = '') {
-            for (const key in data) {
-              if (typeof data[key] === 'object') {
-                parseJson(data[key], `${parentKey}${key}-`);
-              } else {
-                //if data[key] contains a colon, wrap it in quotes
-                if (data[key].includes(':')) {
-                  data[key] = `"${data[key]}"`;
-                }
-                sassVars += `$${parentKey}${key}: ${data[key]};\n`;
-              }
-            }
-          }
-          parseJson(jsonData);
-
-        return {
-            contents: sassVars,
-            syntax: 'scss'
-        }
-    }
-}
-
 const getHelpers = (helpersDir) => {
   let helpers = {}
   fs.readdirSync(helpersDir, { encoding: "utf-8" })
@@ -115,4 +59,4 @@ const getPartials = (partialsDir) => {
 
 
 
-export { getData, getSassData, getFilepaths, getPartials, getHelpers };
+export { getData, getFilepaths, getPartials, getHelpers };
