+++ bin/plugins/postcss/postcss-email-darkmode.js
@@ -6,6 +6,14 @@ const plugin = (opts = {}) => {
       media(atRule) {
         if (atRule.params.includes('prefers-color-scheme: dark')) {
 
+          const otherRules = atRule.params.split('and').filter(rule => !rule.includes('prefers-color-scheme: dark')).map(rule => rule.trim());
+
+          let newAtRule = false;
+
+          if (otherRules.length) {
+            newAtRule = atRule.clone({ params: otherRules.join(' and ') });
+          }
+          
           let prevRootRule = atRule;
 
           atRule.walkRules((rule) => {
@@ -23,11 +31,18 @@ const plugin = (opts = {}) => {
               newRule.selector = `[data-ogsb] ${newRule.selector}`;
             }
 
-            atRule.parent.insertAfter(prevRootRule, newRule);
-
-            prevRootRule = newRule;
-
+            if (newAtRule) {
+              newAtRule.append(newRule);
+            } else {
+              atRule.parent.insertAfter(prevRootRule, newRule);
+              prevRootRule = newRule;
+            }
+            
           });
+
+          if (newAtRule) {
+            atRule.parent.insertAfter(atRule, newAtRule);
+          }
         }
       }
     }
