+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "email-pipeline",
-  "version": "1.1.2",
+  "version": "1.1.3",
   "description": "",
   "main": "bin/index.js",
   "scripts": {
@@ -31,6 +31,7 @@
     "express": "^5.0.0",
     "fs": "^0.0.1-security",
     "glob": "^11.0.0",
+    "globby": "^14.0.2",
     "handlebars": "^4.7.8",
     "html-crush": "^6.0.19",
     "inline-css": "^4.0.2",
