+++ test/src/nunjucks/templates/module.njk
@@ -1,2 +1,4 @@
-<h1>Testing Module2</h1>
+<h1>Testing Module</h1>
 <p>{{base.data}}</p>
+
+<p>test <span style="font-weight: bold;">test</span>.</p>
